Addressed various comments from code review
